requirements.txt
azure-ai-textanalytics==5.1.0
azure-storage-blob
azure-cosmos